Jira issue: SWATCH-2504

## Description
- now uses the new product tag identification methods
  SubscriptionDefinition.getAllProductTagsWithPaygEligibleByRoleOrEngIds
  and

SubscriptionDefinition.getAllProductTagsWithNonPaygEligibleByRoleOrEngIds
- Added unit tests to cover new code

This ensures a common code path for this decision making.


## Testing
Existing unit and iqe tests cover the code changes.

Please test the statement from the Jira card:
"If data comes from telemetry then treat as metered: Y, if from HBI then
metered N, if there is data in both streams then metered Y wins."

### Verification
The successful run of the test suite.
Confirmation of the above statement.

+++ src/main/java/org/candlepin/subscriptions/tally/InventoryAccountUsageCollector.java
@@ -250,13 +250,17 @@ public class InventoryAccountUsageCollector {
         "Reconciling HBI inventoryId={} & swatch inventoryId={}",
         Optional.ofNullable(hbiSystem).map(InventoryHostFacts::getInventoryId),
         Optional.ofNullable(swatchSystem).map(Host::getInventoryId));
+    boolean isMetered = swatchSystem != null && swatchSystem.isMetered();
     if (hbiSystem == null && swatchSystem == null) {
       log.debug("Unexpected, both HBI & Swatch system records are empty");
     } else if (hbiSystem == null) {
-      log.info("Deleting system w/ inventoryId={}", swatchSystem.getInventoryId());
-      hostRepository.delete(swatchSystem);
+      if (!isMetered) {
+        log.info("Deleting system w/ inventoryId={}", swatchSystem.getInventoryId());
+        hostRepository.delete(swatchSystem);
+      }
     } else {
-      NormalizedFacts normalizedFacts = factNormalizer.normalize(hbiSystem, orgHostsData);
+      NormalizedFacts normalizedFacts =
+          factNormalizer.normalize(hbiSystem, orgHostsData, isMetered);
       Set<Key> usageKeys = createHostUsageKeys(applicableProducts, normalizedFacts);
       if (swatchSystem != null) {
         log.debug("Updating system w/ inventoryId={}", hbiSystem.getInventoryId());
